requirements.txt
pytest-homeassistant-custom-component~=0.13.7
flake8~=5.0.4
pytest~=7.2.1
python-fm>=0.0.1
